Apply bugfix for tree_map
This is a bugfix for a bug involving putting an ivy Node object into a
numpy.array. When calling np.array on a tree object (or list containing
a tree object, in this case), the function will hang until a keyboard
interrupt is performed. Specifying the dtype as object fixes this
problem. Possibly this bug stems from something in the code for
np.array, which I think is written in C.